+++ components/planner/index.ts
@@ -14,4 +14,5 @@ export { MiniDailyTimeline } from './MiniDailyTimeline';
 export { ConflictResolutionModal } from './ConflictResolutionModal';
 export { MergedDailyView, type MergedDailyViewRenderOptions } from './MergedDailyView';
 export { SchedulingSidebar } from './SchedulingSidebar';
+export { QuickPlanDayModal, type QuickPlanDraftTask } from './modals/QuickPlanDayModal';